+++ csrc/kernels/deep_ep/intranode.cu
@@ -20,7 +20,7 @@ namespace intranode {
 template <int kNumRanks>
 __global__ void
 notify_dispatch(const int *num_tokens_per_rank, int *moe_recv_counter_mapped,
-                const int *num_tokens_per_expert, int *moe_recv_expert_counter_mapped,
+                const int *num_tokens_per_expert, int64_t *moe_recv_expert_counter_mapped,
                 int64_t *moe_recv_tokens_per_experts, int num_experts, int num_tokens,
                 int num_channels, const bool *is_token_in_rank, int *channel_prefix_matrix,
                 int *rank_prefix_matrix_copy, int num_memset_int, int expert_alignment,
